requirements.txt
instagrapi>=2.1.0
mcp>=1.0.0
playwright>=1.40.0
playwright-stealth>=1.0.0
googlesearch-python>=1.2.0
